app/requirements.txt
Flask==2.0
flask_restful==0.3.9
pyspark==3.2.1
python_crontab==2.6.0
requests==2.27.1
pymonetdb==1.5.1
